+++ experimental/packages/opentelemetry-sdk-node/src/sdk.ts
@@ -24,7 +24,7 @@ import {
 } from '@opentelemetry/instrumentation';
 import { NodeTracerConfig, NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import {
-  detectResources,
+  detectResourcesSync,
   envDetector,
   processDetector,
   Resource,
@@ -113,13 +113,13 @@ export class NodeSDK {
   }
 
   /** Detect resource attributes */
-  public async detectResources(config?: ResourceDetectionConfig): Promise<void> {
+  public detectResources(config?: ResourceDetectionConfig): void {
     const internalConfig: ResourceDetectionConfig = {
       detectors: [ envDetector, processDetector],
       ...config,
     };
 
-    this.addResource(await detectResources(internalConfig));
+    this.addResource(detectResourcesSync(internalConfig));
   }
 
   /** Manually add a resource */
@@ -130,9 +130,9 @@ export class NodeSDK {
   /**
    * Once the SDK has been configured, call this method to construct SDK components and register them with the OpenTelemetry API.
    */
-  public async start(): Promise<void> {
+  public start(): void {
     if (this._autoDetectResources) {
-      await this.detectResources();
+      this.detectResources();
     }
 
     if (this._tracerProviderConfig) {