+++ packages/opentelemetry-sdk-trace-base/src/export/BatchSpanProcessorBase.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { context, Context, TraceFlags } from '@opentelemetry/api';
+import { context, Context, diag, TraceFlags } from '@opentelemetry/api';
 import {
   BindOnceFuture,
   ExportResultCode,
@@ -152,8 +152,10 @@ export abstract class BatchSpanProcessorBase<T extends BufferConfig> implements 
         // Reset the finished spans buffer here because the next invocations of the _flush method
         // could pass the same finished spans to the exporter if the buffer is cleared
         // outside of the execution of this callback.
-        this._exporter.export(
-          this._finishedSpans.splice(0, this._maxExportBatchSize),
+        const spans = this._finishedSpans.splice(0, this._maxExportBatchSize);
+
+        const doExport = () => this._exporter.export(
+          spans,
           result => {
             clearTimeout(timer);
             if (result.code === ExportResultCode.SUCCESS) {
@@ -166,6 +168,17 @@ export abstract class BatchSpanProcessorBase<T extends BufferConfig> implements 
             }
           }
         );
+
+        const pendingResources = spans.map(span => span.resource)
+          .filter(resource => !resource.asyncAttributesHaveResolved());
+
+        // Avoid scheduling a promise to make the behavior more predictable and easier to test
+        if (pendingResources.length === 0) {
+          doExport();
+        } else {
+          Promise.all(pendingResources.map(resource => resource.waitForAsyncAttributes()))
+            .then(doExport, err => diag.debug('Error while resolving async portion of resource: ', err));
+        }
       });
     });
   }